server/requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.23
alembic==1.13.1
sqlite3

# Validation et sérialisation
email-validator==2.1.0
python-multipart==0.0.6

# Logging et monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Développement et tests
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
coverage==7.3.2

# Utilitaires
python-dotenv==1.0.0
typer==0.9.0
click==8.1.7

# Images (pour intégration avec le core C++)
pillow==10.1.0
numpy==1.25.2
